Updated switch statement and loops w/ setElem()
Uses the setElem() method in nested for loops to get all the int values from the file, they are then put in a 2d array. Switch statement is now more efficient as it only sets isError to true if the user enters a wrong answer (earlier I set isError to false in every single case, which looked dumb). It also exits the program if the user enters 'q' and will otherwise loop repeatedly.